chore(Gradle) Detekt and KTlint basic setup
Integrating Detekt and KTlint in their very basic setup. Due to issues with .editorconfig and KTlint, editorconfig was removed for now, as KTlint will execute the same checks.

The max line length for detekt checks was changed from the default of 120 to 200, as most devs today use widescreen anyway.

Issues with detekt and KTlint where not fixed in this commit.