chore(All): Renaming and domain module
Introducing a domain module as an example to showcase where to put domain logic. There can be multiple domain modules if required in order to separate different concerns.

Also renaming the application module to infrastructure, to indicate clearly what the purpose of the module is. What's missing so far, is a commons module, serving utilities.