feat(Infrastructure): Optional config
Implementing the possibility to have optional config values and act on them. In the given example, the database configuration is optional, and if we try to load it with the given mechanics, we have the possibility to just not connect to the database if the values are not present.

Right now, those mechanics are implemented on a very basic level, and might need further improvements over time.